+++ src/components/MemoriWidget/MemoriWidget.stories.tsx
@@ -262,3 +262,5 @@ WithUserAvatarAsElement.args = {
   tenant,
   userAvatar: <span>USER</span>,
 };
+
+