+++ src/components/layouts/layouts.stories.tsx
@@ -50,6 +50,7 @@ DefaultLayout.args = {
   showUpload: true,
   showReasoning: false,
   showLogin: true,
+  multilingual: true,
 };